Drop dependency on gitlib
Gitlib does not have support for modifying the Git config, and either
way, libgit2 has no functionality to find a branch point merge base. So
I'll resort to invoking Git commands and parsing their output. The
things I want to do are not terribly complex anyway, so this should be
fine.

Plus, I get a lot of error handling for free: Git will print its
standard error message if the working directory is not a repository for
instance, and all I need to do is bail out if Git returns a nonzero exit
code.