Add functions that call Git
This reduces the commands supported by the free monad to two: one for
calling Git, and one for doing a regular IO operation. All other
functionality can be implemented in terms of this, by making the right
calls, and then parsing the output. A first step in this direction is
the listBranches function, which I implemented just to test for now.