Add graph data structure
This will be used to construct and manipulate the dependency graph.
There is a Data.Graph module in the containers package, but its vertices
are indexed. For my use case something simpler is sufficient, and
performance is not really an issue. In practice Git repositories have a
few dozen branches at most, and if you want to have dependencies between
more branches than that, you have a bigger issue than performance.